Split workspace activity automation surface tests
