+++ Tests/QuillCodeAppTests/WorkspaceSurfaceTests.swift
@@ -1,5 +1,4 @@
 import XCTest
-import QuillCodeAgent
 import QuillCodeCore
 import QuillCodeTools
 import QuillComputerUseKit
@@ -282,304 +281,6 @@ final class WorkspaceSurfaceTests: XCTestCase {
         XCTAssertEqual(surface.sidebar.items.count, 0)
         XCTAssertNil(surface.sidebar.selectedThreadID)
         XCTAssertFalse(surface.sidebar.isSelectionMode)
-    }
-
-    func testActivitySurfaceSummarizesThreadToolsSourcesAndArtifacts() throws {
-        let instruction = ProjectInstruction(
-            path: ".quillcode/AGENTS.md",
-            title: "AGENTS.md",
-            content: "Use the repo patterns.",
-            byteCount: 22
-        )
-        let memory = MemoryNote(
-            id: "global-note",
-            scope: .global,
-            title: "Prefers concise diffs",
-            content: "Keep changes reviewable.",
-            relativePath: "preferences.md",
-            byteCount: 24
-        )
-        let call = ToolCall(
-            id: "tool-activity",
-            name: ToolDefinition.shellRun.name,
-            argumentsJSON: #"{"cmd":"whoami"}"#
-        )
-        let result = ToolResult(
-            ok: true,
-            stdout: "quill\n",
-            artifacts: ["/tmp/quillcode-activity.png"]
-        )
-        let thread = ChatThread(
-            title: "Run command",
-            messages: [
-                .init(role: .user, content: "run whoami"),
-                .init(role: .assistant, content: "Output:\nquill")
-            ],
-            events: [
-                .init(kind: .message, summary: "run whoami"),
-                .init(kind: .toolQueued, summary: "host.shell.run queued", payloadJSON: try JSONHelpers.encodePretty(call)),
-                .init(kind: .toolRunning, summary: "host.shell.run running"),
-                .init(kind: .toolCompleted, summary: "host.shell.run completed", payloadJSON: try JSONHelpers.encodePretty(result)),
-                .init(kind: .message, summary: "Output:\nquill")
-            ],
-            instructions: [instruction],
-            memories: [memory]
-        )
-        let model = QuillCodeWorkspaceModel(
-            root: QuillCodeRootState(
-                threads: [thread],
-                selectedThreadID: thread.id
-            ),
-            activity: ActivityState(isVisible: true)
-        )
-
-        let activity = model.surface().activity
-
-        XCTAssertTrue(activity.isVisible)
-        XCTAssertEqual(activity.taskTitle, "run whoami")
-        XCTAssertEqual(activity.tools.map(\.title), [ToolDefinition.shellRun.name])
-        XCTAssertEqual(activity.tools.first?.statusLabel, ToolCardStatus.done.rawValue)
-        XCTAssertEqual(activity.artifacts.map(\.label), ["quillcode-activity.png"])
-        XCTAssertEqual(activity.sources.map(\.title), ["AGENTS.md", "Prefers concise diffs"])
-        XCTAssertEqual(activity.finalAnswer, "Output: quill")
-        XCTAssertEqual(activity.planItems.map(\.title), [
-            "Understand request",
-            "Load context",
-            "Use tools",
-            "Review results",
-            "Answer user"
-        ])
-        XCTAssertEqual(activity.planItems.map(\.statusLabel), ["Done", "Done", "Done", "Done", "Done"])
-        XCTAssertTrue(activity.planItems.contains { $0.title == "Use tools" && $0.detail.contains(ToolDefinition.shellRun.name) })
-        XCTAssertTrue(activity.handoffSummary?.contains("Thread: Run command") == true)
-        XCTAssertTrue(activity.handoffSummary?.contains("Latest request: run whoami") == true)
-        XCTAssertTrue(activity.handoffSummary?.contains("Tools: 1 tool (\(ToolDefinition.shellRun.name))") == true)
-        XCTAssertTrue(activity.handoffSummary?.contains("Artifacts: 1 artifact (quillcode-activity.png)") == true)
-        XCTAssertTrue(activity.recentSteps.contains { $0.title == "Tool completed" && $0.statusLabel == "Done" })
-        XCTAssertEqual(activity.sections.map(\.kind), [.plan, .recent, .handoff, .tools, .sources, .artifacts, .latestAnswer])
-        XCTAssertEqual(activity.sections.first { $0.kind == .plan }?.items.map(\.title), activity.planItems.map(\.title))
-        XCTAssertEqual(activity.sections.first { $0.kind == .plan }?.countLabel, "5 items")
-        XCTAssertEqual(activity.sections.first { $0.kind == .handoff }?.bodyText, activity.handoffSummary)
-        XCTAssertEqual(activity.sections.first { $0.kind == .handoff }?.countLabel, "1 summary")
-        XCTAssertEqual(activity.sections.first { $0.kind == .tools }?.items.map(\.title), [ToolDefinition.shellRun.name])
-        XCTAssertEqual(activity.sections.first { $0.kind == .artifacts }?.artifacts.map(\.label), ["quillcode-activity.png"])
-        XCTAssertEqual(activity.sections.first { $0.kind == .latestAnswer }?.bodyText, "Output: quill")
-        XCTAssertEqual(activity.sections.first { $0.kind == .tools }?.toggleCommandID, "activity-toggle-section:tools")
-    }
-
-    func testActivitySurfacePrefersModelAuthoredPlan() throws {
-        let update = AgentPlanUpdate(
-            explanation: "The model is planning the work directly.",
-            plan: [
-                AgentPlanItem(step: "Inspect current state", status: .completed),
-                AgentPlanItem(step: "Apply focused change", status: .inProgress, detail: "Keep the diff small."),
-                AgentPlanItem(step: "Run validation", status: .pending)
-            ]
-        )
-        let result = ToolResult(ok: true, stdout: try JSONHelpers.encodePretty(update))
-        let thread = ChatThread(
-            title: "Plan work",
-            messages: [.init(role: .user, content: "plan the work")],
-            events: [
-                .init(
-                    kind: .toolCompleted,
-                    summary: "\(ToolDefinition.planUpdate.name) completed",
-                    payloadJSON: try JSONHelpers.encodePretty(result)
-                )
-            ]
-        )
-        let model = QuillCodeWorkspaceModel(
-            root: QuillCodeRootState(threads: [thread], selectedThreadID: thread.id),
-            activity: ActivityState(isVisible: true)
-        )
-
-        let activity = model.surface().activity
-
-        XCTAssertEqual(activity.planItems.map(\.title), [
-            "Inspect current state",
-            "Apply focused change",
-            "Run validation"
-        ])
-        XCTAssertEqual(activity.planItems.map(\.statusLabel), ["Done", "Running", "Pending"])
-        XCTAssertEqual(activity.planItems[0].detail, "The model is planning the work directly.")
-        XCTAssertEqual(activity.planItems[1].detail, "Keep the diff small.")
-        XCTAssertEqual(activity.sections.first { $0.kind == .plan }?.countLabel, "3 items")
-        XCTAssertEqual(activity.sections.first { $0.kind == .plan }?.items.map(\.kind), [
-            "authored-plan",
-            "authored-plan",
-            "authored-plan"
-        ])
-    }
-
-    func testActivityCommandTogglesActivityPane() {
-        let model = QuillCodeWorkspaceModel()
-
-        XCTAssertFalse(model.surface().activity.isVisible)
-        XCTAssertTrue(model.runWorkspaceCommand("toggle-activity", workspaceRoot: URL(fileURLWithPath: "/tmp")))
-        XCTAssertTrue(model.surface().activity.isVisible)
-    }
-
-    func testAutomationsCommandTogglesAutomationsPaneWithoutActivity() {
-        let model = QuillCodeWorkspaceModel()
-
-        XCTAssertFalse(model.surface().automations.isVisible)
-        XCTAssertFalse(model.surface().activity.isVisible)
-        XCTAssertTrue(model.runWorkspaceCommand("toggle-automations", workspaceRoot: URL(fileURLWithPath: "/tmp")))
-
-        let surface = model.surface()
-        XCTAssertTrue(surface.automations.isVisible)
-        XCTAssertFalse(surface.activity.isVisible)
-        XCTAssertEqual(surface.automations.title, "Automations")
-        XCTAssertEqual(surface.automations.workflows.map(\.title), [
-            "Thread follow-ups",
-            "Workspace schedules",
-            "Monitors"
-        ])
-    }
-
-    func testAutomationsSurfaceUsesConfiguredAutomationRowsWhenPresent() {
-        let model = QuillCodeWorkspaceModel(automations: AutomationsState(items: [
-            QuillAutomation(
-                title: "Nightly repo check",
-                detail: "Run tests and summarize failures.",
-                kind: .workspaceSchedule,
-                scheduleKind: .cron,
-                scheduleDescription: "Every weekday at 6:00 PM"
-            ),
-            QuillAutomation(
-                title: "Paused PR monitor",
-                detail: "Watch the launch PR after review starts.",
-                kind: .monitor,
-                status: .paused,
-                scheduleKind: .event,
-                scheduleDescription: "PR events"
-            )
-        ]))
-
-        let automations = model.surface().automations
-
-        XCTAssertEqual(automations.statusLabel, "1 active · 1 paused")
-        XCTAssertEqual(automations.workflows.map(\.title), ["Nightly repo check", "Paused PR monitor"])
-        XCTAssertEqual(automations.workflows.map(\.statusLabel), ["Active", "Paused"])
-        XCTAssertEqual(automations.workflows.first?.scheduleLabel, "Every weekday at 6:00 PM")
-        XCTAssertEqual(automations.workflows.first?.runActionTitle, "Run now")
-        XCTAssertTrue(automations.workflows.first?.runCommandID?.hasPrefix("automation-run:") == true)
-        XCTAssertEqual(automations.workflows.first?.primaryActionTitle, "Pause")
-        XCTAssertTrue(automations.workflows.first?.primaryCommandID?.hasPrefix("automation-pause:") == true)
-        XCTAssertTrue(automations.workflows.first?.deleteCommandID?.hasPrefix("automation-delete:") == true)
-        XCTAssertEqual(automations.workflows.last?.primaryActionTitle, "Resume")
-        XCTAssertTrue(automations.workflows.last?.primaryCommandID?.hasPrefix("automation-resume:") == true)
-    }
-
-    func testThreadFollowUpAutomationsExposeRunNowAction() {
-        let model = QuillCodeWorkspaceModel(automations: AutomationsState(items: [
-            QuillAutomation(
-                title: "Launch follow-up",
-                detail: "Resume the launch thread.",
-                kind: .threadFollowUp,
-                scheduleKind: .heartbeat,
-                scheduleDescription: "Manual follow-up",
-                threadID: UUID()
-            ),
-            QuillAutomation(
-                title: "Paused follow-up",
-                detail: "Resume later.",
-                kind: .threadFollowUp,
-                status: .paused,
-                scheduleKind: .heartbeat,
-                scheduleDescription: "Manual follow-up",
-                threadID: UUID()
-            )
-        ]))
-
-        let automations = model.surface().automations
-
-        XCTAssertEqual(automations.workflows.first?.runActionTitle, "Run now")
-        XCTAssertTrue(automations.workflows.first?.runCommandID?.hasPrefix("automation-run:") == true)
-        XCTAssertNil(automations.workflows.last?.runActionTitle)
-        XCTAssertNil(automations.workflows.last?.runCommandID)
-    }
-
-    func testAutomationsSurfaceExposesCreateCommandsForSelectedThreadAndProject() {
-        let project = ProjectRef(name: "QuillCode", path: "/tmp/QuillCode")
-        let thread = ChatThread(title: "Ship QuillCode", projectID: project.id)
-        let model = QuillCodeWorkspaceModel(root: QuillCodeRootState(
-            projects: [project],
-            selectedProjectID: project.id,
-            threads: [thread],
-            selectedThreadID: thread.id
-        ))
-
-        let automations = model.surface().automations
-
-        XCTAssertEqual(automations.createThreadFollowUpCommand?.id, "automation-create-thread-follow-up")
-        XCTAssertEqual(automations.createThreadFollowUpCommand?.category, WorkspaceCommandPalette.automationsCategory)
-        XCTAssertEqual(automations.createThreadFollowUpCommand?.isEnabled, true)
-        XCTAssertEqual(automations.scheduleThreadFollowUpCommands.map(\.id), [
-            "automation-create-thread-follow-up-after:600",
-            "automation-create-thread-follow-up-after:3600",
-            "automation-create-thread-follow-up-tomorrow",
-            "automation-create-thread-follow-up-every:daily"
-        ])
-        XCTAssertEqual(automations.scheduleThreadFollowUpCommands.map(\.isEnabled), [true, true, true, true])
-        XCTAssertEqual(model.surface().commands.first { $0.id == "automation-create-thread-follow-up" }?.isEnabled, true)
-        XCTAssertEqual(
-            model.surface().commands
-                .filter { $0.id.hasPrefix("automation-create-thread-follow-up-after:") }
-                .map(\.isEnabled),
-            [true, true]
-        )
-        XCTAssertEqual(automations.createWorkspaceScheduleCommand?.id, "automation-create-workspace-schedule")
-        XCTAssertEqual(automations.createWorkspaceScheduleCommand?.category, WorkspaceCommandPalette.automationsCategory)
-        XCTAssertEqual(automations.createWorkspaceScheduleCommand?.isEnabled, true)
-        XCTAssertEqual(automations.scheduleWorkspaceScheduleCommands.map(\.id), [
-            "automation-create-workspace-schedule-after:600",
-            "automation-create-workspace-schedule-after:3600",
-            "automation-create-workspace-schedule-tomorrow",
-            "automation-create-workspace-schedule-every:daily"
-        ])
-        XCTAssertEqual(automations.scheduleWorkspaceScheduleCommands.map(\.isEnabled), [true, true, true, true])
-        XCTAssertEqual(model.surface().commands.first { $0.id == "automation-create-workspace-schedule" }?.isEnabled, true)
-        XCTAssertEqual(
-            model.surface().commands
-                .filter { $0.id.hasPrefix("automation-create-workspace-schedule-after:") }
-                .map(\.isEnabled),
-            [true, true]
-        )
-        XCTAssertEqual(model.surface().commands.first { $0.id == "automation-create-workspace-schedule-tomorrow" }?.isEnabled, true)
-    }
-
-    func testActivitySectionToggleCollapsesSharedSurfaceSection() throws {
-        let call = ToolCall(
-            id: "tool-activity",
-            name: ToolDefinition.shellRun.name,
-            argumentsJSON: #"{"cmd":"whoami"}"#
-        )
-        let result = ToolResult(ok: true, stdout: "quill\n")
-        let thread = ChatThread(
-            title: "Run command",
-            messages: [.init(role: .user, content: "run whoami")],
-            events: [
-                .init(kind: .toolQueued, summary: "host.shell.run queued", payloadJSON: try JSONHelpers.encodePretty(call)),
-                .init(kind: .toolCompleted, summary: "host.shell.run completed", payloadJSON: try JSONHelpers.encodePretty(result))
-            ]
-        )
-        let model = QuillCodeWorkspaceModel(
-            root: QuillCodeRootState(threads: [thread], selectedThreadID: thread.id),
-            activity: ActivityState(isVisible: true)
-        )
-
-        XCTAssertEqual(model.surface().activity.sections.first { $0.kind == .tools }?.isCollapsed, false)
-        XCTAssertTrue(model.runWorkspaceCommand("activity-toggle-section:tools", workspaceRoot: URL(fileURLWithPath: "/tmp")))
-        XCTAssertEqual(model.surface().activity.sections.first { $0.kind == .tools }?.isCollapsed, true)
-        XCTAssertTrue(model.surface().activity.isVisible)
-        XCTAssertTrue(model.runWorkspaceCommand("activity-toggle-section:handoff", workspaceRoot: URL(fileURLWithPath: "/tmp")))
-        XCTAssertEqual(model.surface().activity.sections.first { $0.kind == .handoff }?.isCollapsed, true)
-        XCTAssertTrue(model.runWorkspaceCommand("activity-toggle-section:plan", workspaceRoot: URL(fileURLWithPath: "/tmp")))
-        XCTAssertEqual(model.surface().activity.sections.first { $0.kind == .plan }?.isCollapsed, true)
-        XCTAssertTrue(model.runWorkspaceCommand("activity-toggle-section:tools", workspaceRoot: URL(fileURLWithPath: "/tmp")))
-        XCTAssertEqual(model.surface().activity.sections.first { $0.kind == .tools }?.isCollapsed, false)
-        XCTAssertFalse(model.runWorkspaceCommand("activity-toggle-section:not-real", workspaceRoot: URL(fileURLWithPath: "/tmp")))
     }
 
     func testSidebarSearchExcludesHiddenToolFeedback() {
